requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
